Revert "userspace, libc: try to use stdint types"
This reverts commit 37c091cf20ebf6a0fd516aa8e026af1d5af30382.

+++ pkgs/gfx/cairodem.c
@@ -1,5 +1,4 @@
 #include <cairo/cairo.h>
-#include <stdint.h>
 #include <stdio.h>
 #include <sys/gfx.h>
 #include <sys/ioctl.h>
@@ -23,7 +22,7 @@ int main(int argc, char const **argv)
 
     struct fbdev_bitblit bitblit = {
         .target_buffer = GFX_FRONT_BUFFER,
-        .source = (uint32_t *)cairo_image_surface_get_data(surface),
+        .source = (unsigned long*)cairo_image_surface_get_data(surface),
         .x = 0,
         .y = 0,
         .width = WIDTH,
